requirements.txt
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-llms-anthropic>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.0
nest-asyncio>=1.5.0
pypdf>=4.0.0
